Ref: ParticipantFactory 코드를 파일로 분리
